sw: inside: use MQ135 lib to read co2 sensor values
This patch only makes use of the lib but the values are
not correct yet as we need to calibrate the sensor first.
This calibration will be part of a later commit which
will allow calibration from the settings UI mode.

This patch also change the reference voltage for the analog
read to 5V as this is the output range of the sensor

Signed-off-by: Pierre-Yves Kerbrat <pkerbrat@free.fr>